increase bf16 gemm test tolerance

+++ example/01_gemm/gemm_wmma_int8.cpp
@@ -8,7 +8,7 @@
 using ADataType        = int8_t;
 using BDataType        = int8_t;
 using AccDataType      = int32_t;
-using CShuffleDataType = int8_t;
+using CShuffleDataType = int32_t;
 using CDataType        = int8_t;
 
 using ALayout = Row;
